fix(review): address PR #915 review comments
- Fix speed row UX: show only available directions instead of '--'
- Extract shared MetricTile to layout_blocks (remove duplication)
- Refactor NetworkBadgeWidget to use AppBadge from UI Kit
  (icon support pending #916)
